quiz_gen_django/requirements.txt
Django>=5.1.3
python-dotenv>=1.0.1
wikipedia>=1.4.0
spacy==3.8.2
sentence-transformers==3.3.1
sentencepiece==0.2.0
transformers==4.46.2
httpx==0.28.1
llama_cpp_python>=0.3.4
ru_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/ru_core_news_sm-3.8.0/ru_core_news_sm-3.8.0-py3-none-any.whl
-e .